soc/requirements.txt
clearbit==0.1.7
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Markdown==3.3.3
psycopg2-binary>=2.8
pyhunter==1.7
